(cleanup, nobug) Start BVC split work, splitting out Nav Delegate and KeyCommands These were obvious low-hanging fruit items in regards to splitting up this file. There is much more to go.
